requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
requests
pydantic
python-dotenv
mangum

